+++ tasks/helpers/OptionHelper.js
@@ -1,0 +1,7 @@
+let options = {
+    imageResize: true,
+    fontGeneration: true,
+    deepCleanSrc: false,
+};
+
+exports.options = options;