+++ tasks/helpers/VariableHelper.js
@@ -1,0 +1,7 @@
+let baseDir = "src"; // If you wanna change you main dev directory, just replace src to another name
+let deployDir = "public";
+let imagesWatch = "*.jpeg, *.jpg, *.webp, *.png, *.svg";
+
+exports.baseDir = baseDir;
+exports.deployDir = deployDir;
+exports.imagesWatch = imagesWatch;