+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+### [0.0.3](https://github.com/waveofdandelion/dandelion-gpack/compare/v0.0.2...v0.0.3) (2022-09-24)
+
+
+### Features
+
+* add gulpfile.js to the project ([4241c5e](https://github.com/waveofdandelion/dandelion-gpack/commit/4241c5e9d22901cf86cde4be28d76e8126b351fb))
+* **helpers:** add options for users like deepClean and imageResize ([b253706](https://github.com/waveofdandelion/dandelion-gpack/commit/b253706ec10b0275709629364ea2faf3fa5137d0))
+* **helpers:** add predefined paths for configs ([534ddaa](https://github.com/waveofdandelion/dandelion-gpack/commit/534ddaa06a22ef77950957b6d0006e0b1270c950))
+* **helpers:** add predefined variables for dist and src ([c14acc7](https://github.com/waveofdandelion/dandelion-gpack/commit/c14acc782b81944310e4ec65b94f9648c1ce457c))
+
+
+### Docs
+
+* update readme ([550822d](https://github.com/waveofdandelion/dandelion-gpack/commit/550822d762ead64425b2bec3a1308ebef6064374))
+
 ### [0.0.2](https://github.com/waveofdandelion/dandelion-gpack/compare/v0.0.1...v0.0.2) (2022-09-18)